fix: feed Bresenham fractional position into controller to damp low-speed vel-fb oscillation
At sub-1-step-per-period velocities the Bresenham accumulator fires 0 or 1
step per period.  The position controller saw only the integer abs_pos_achieved
and would oscillate velocity_q when the quantised position briefly crossed the
1-step dead zone, causing audible unevenness and a sawtooth vel-fb signal.

Add the accumulated-but-unfired fractional steps to abs_pos_achieved before
passing to compute_velocity_cmd.  The controller now sees sub-step position,
keeping the correction inside the dead zone throughout the Bresenham cycle.

Changes:
- compute_velocity_cmd: abs_pos_achieved int32_t → double (enables fractional pos)
- do_steps: compute pos_frac = abs_pos_achieved + sign * step_accumulator_q/65536
- Two new tests: direct API test for fractional dead zone suppression; integration
  test that builds accumulator over 3 periods then verifies correction is damped

Closes #39.

+++ src/rp2040/pio.h
@@ -37,7 +37,7 @@ double compute_velocity_cmd(
     uint8_t  cmd_type,
     double   velocity_requested,
     double   abs_pos_requested,
-    int32_t  abs_pos_achieved,
+    double   abs_pos_achieved,
     uint8_t  enabled,
     uint32_t updated,
     uint32_t update_period_us,
